Fix data type error of pandas series and add quick snap
After python39, pd.series set object as data default type rather than float, pass a dtype parameter to silence this warning, then add bond index and stock index to import quick snap.

+++ RiskQuantLib/CompanyList/ListedCompanyList/listedCompanyList.py
@@ -22,7 +22,7 @@ class listedCompanyList(baseList,setListedCompanyList):
         tmpList = self.all + [listedCompany(codeString, nameString, companyTypeString)]
         self.setAll(tmpList)
 
-    def addListedCompanySeries(self, listedCompanyCodeSeries, listedCompanyNameSeries = pd.Series(), companyTypeString = 'Listed Company'):
+    def addListedCompanySeries(self, listedCompanyCodeSeries, listedCompanyNameSeries = pd.Series(dtype=float), companyTypeString = 'Listed Company'):
         """
         Add lots of companies into list.
         """
